% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precursors.R
\name{precursors}
\alias{precursors}
\title{Filter precursor data}
\usage{
precursors(diann_data, proteotypic = TRUE, remove_contaminants = FALSE)
}
\arguments{
\item{diann_data}{DiaNNData R6 object (not DiaNNData)}

\item{proteotypic}{logical. Filter for proteotypic peptides only (default: TRUE)}

\item{remove_contaminants}{logical. Remove contaminant proteins (default: FALSE)}
}
\value{
data.frame. Filtered precursor data
}
\description{
Filter precursor data
}
